fix(push): harden presence gating (self-review)
Address self-review findings on the interaction-gated suppression:

- isDevicePresent/recordInteractionPresence now swallow Cache Storage
  errors (private mode, quota). The SW read runs inside event.waitUntil,
  so an unguarded throw would reject the handler and silently drop the
  notification — the opposite of the intended fail-open. On error we
  treat the device as not-present and show the notification.
- Register the "threa-presence" cache in the SW activate cleanup
  allowlist so it isn't wiped on every deploy.
- Hoist the 2-minute interaction window to @threa/types
  (PRESENCE_INTERACTION_WINDOW_MS), shared by the backend attended-device
  routing and the frontend SW so the two layers cannot drift.
- Extract the presence writer from SocketProvider into a dedicated
  useSwPresence hook (matches the useAutoMarkAsRead pattern), reusing the
  existing page-interaction tracker.
